+++ src/places/types/place-boundaries.type.ts
@@ -1,0 +1,9 @@
+export type Coordinates = {
+  lat: number;
+  long: number;
+};
+
+export type PlaceBoundaries = {
+  topLeft: Coordinates;
+  bottomRight: Coordinates;
+};
