--- src/places/types/placeBoundaries.ts
@@ -1,3 +1,0 @@
-export type Coordinates = { lat: number; long: number };
-
-export type PlaceBoundaries = { topLeft: Coordinates; bottomRight: Coordinates };
